Move processors to custom packages
